Moved UnZip to kura.util
