+++ kura/org.eclipse.kura.core.cloud/META-INF/MANIFEST.MF
@@ -40,6 +40,7 @@ Import-Package: com.eclipsesource.json;version="0.9.5",
  org.eclipse.kura.position;version="[1.0,2.0)",
  org.eclipse.kura.security.tamper.detection;version="[1.0,2.0)",
  org.eclipse.kura.system;version="[1.5,2.0)",
+ org.eclipse.kura.util.zip;version="[1.0,2.0)",
  org.osgi.framework;version="1.5.0",
  org.osgi.service.component;version="1.2.0",
  org.osgi.service.event;version="1.3.0",
