+++ kura/org.eclipse.kura.util/META-INF/MANIFEST.MF
@@ -17,5 +17,6 @@ Export-Package: org.eclipse.kura.util.base;version="1.2.0",
  org.eclipse.kura.util.collection;version="1.0.0",
  org.eclipse.kura.util.configuration;version="1.0.0",
  org.eclipse.kura.util.osgi;version="1.0.0",
- org.eclipse.kura.util.service;version="1.1.0"
+ org.eclipse.kura.util.service;version="1.1.0",
+ org.eclipse.kura.util.zip;version="1.0.0"
 Bundle-ActivationPolicy: lazy
